+++ lucene/sandbox/src/java/org/apache/lucene/sandbox/facet/plain/histograms/HistogramCollector.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.facet.histogram;
+package org.apache.lucene.sandbox.facet.plain.histograms;
 
 import java.io.IOException;
 import org.apache.lucene.index.DocValues;
